Model execution, rewind and replay changes and fixes

+++ common/src/main/java/com/fyntrac/common/entity/AccountBalance.java
@@ -9,6 +9,7 @@ import org.springframework.data.mongodb.core.index.Indexed;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @SuperBuilder
@@ -24,7 +25,7 @@ public class AccountBalance implements Serializable {
     protected AccountType accountType;
     protected String transactionName;
     protected int periodId;
-    protected double amount;
+    protected BigDecimal amount;
     protected String accountNumber;
     protected String accountName;
     protected String accountSubtype;
